Fix path encoding for special characters in files
This commit makes it so that paths always are the path
they are expected to be.  In addition it fixes a couple
times when PageJS was forcing a full page load to occur
by prefixing the paths properly.

This commit also fixes the arrows which regressed as
well during the change to PageJS

fixes CNVS-28916

Test Plan:
  - Go to files
  - Create a folder called 'this#could+be bad? maybe'
  - Upload a file into it
  - The file preview should open and close properly
    when attempting to open the file.
  - Additionally, it should not have performed a full
    page reload when opening or closing the file.

Change-Id: I2fe62d32d2efc124c6dad614edfd3236a42967bb
Reviewed-on: https://gerrit.instructure.com/77869
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>
Product-Review: Clay Diffrient <cdiffrient@instructure.com>