Fix yard-related errors/warnings
Yard is having trouble with multi-line statements that
use dot-first notation and have comments at the end of line.
The commit moves to end-of-line dot notation (which is
more standard in the codebase) and removes some obvious
comments. Either change would resolve the issue.

This also switch one @param tag to @argument since it is
not a param.

Test Plan:
  - Run 'rake doc:api'
  - The Yard docs should be built without errors/warnings