Enable "Opt Out of Inbox" feature for NUI
Fixes: CNVS-29400

Users should be able to opt out of the
inbox notification badge if the feature
is enabled. It was not working in the
NUI, and this commit fixes it.

Test Plan:

- In the NUI and as an admin, go to
'Feature Options', and enable the 'Allow
Users to Opt-out of the Inbox' feature
http://www.screencast.com/t/6kbam6kxV

- Send a message to a course participant

- Switch to that participant and go to
Account -> Settings

- You should see a notification badge
on the 'Inbox' icon on the global nav
sidebar indicating that you received
a message

- Select the checkbox under the
'Conversations Inbox Opt-out' to
disable the notificaton badge
http://www.screencast.com/t/lnCGBtOLg

- Refresh the page and you should no
longer see a notification badge on
the 'Inbox' icon on the global nav
sidebar.

Change-Id: I178cc4a8e0fc889ad8fb8f2e95f8bc6989d949b7
Reviewed-on: https://gerrit.instructure.com/80665
Reviewed-by: Simon Williams <simon@instructure.com>
Reviewed-by: Ryan Shaw <ryan@instructure.com>
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
Tested-by: Jenkins
QA-Review: Myller de Araujo <myller@instructure.com>
Product-Review: Colleen Palmer <colleen@instructure.com>