Fix theme editor filename upload overflow
Fixes: CNVS-29335

Currently when uploading a custom
css or javascript file in theme
editor, if the filename is too long
the form input expands and is cut off.
http://www.screencast.com/t/4EKhoSsYn0v
This commit prevents form cutoff and
adds an ellipsis if the filename is
too long.

Test Plan:

- Go to theme editor

- Upload custom css file and a javascript
file with lengthy names

- The form input size should not expand
with the filename length, but should be
consistent.

- Additionally, the filename text is
truncated with an ellipsis

Change-Id: I5a2c7c89b15581c9d344672152f4ad7b03df4f8d
Reviewed-on: https://gerrit.instructure.com/81008
Reviewed-by: Pam Hiett <phiett@instructure.com>
Tested-by: Jenkins
QA-Review: Myller de Araujo <myller@instructure.com>
Product-Review: Colleen Palmer <colleen@instructure.com>