fix API links in headers
fixes CNVS-29618

The API pagination is supposed to put links in the headers
when responding to a request. Previously the polls controllers
along with anything still using the jsonapi pagination only
included the links inside the meta tag. Now the links are in
the headers as well, that way none of the api consumers will be
broken.

test plan:
1. Create some polls and poll sessions, and get it all hooked up
2. Make sure to have an access token to allow authorization for the
   API requests.
3. Using an API request (i used postman) test the links, make sure
   to have the per_page=1 parameter passed through the url
4. Inspect the headers in the response and notice that the links
   now there.
5. You may also want to pass 'Accept:application/vnd.api+json' in
   the headers alongside the authorization header. This will force
   the use of the old jsonapi pagination.
6. In a browser, open the individual links from the response headers
   and make sure the correct information is displayed.

Change-Id: I0df0bcc4b1f332b8a99885c57154865258db19d3
Reviewed-on: https://gerrit.instructure.com/82221
Reviewed-by: Bryan Petty <bpetty@instructure.com>
Tested-by: Jenkins
QA-Review: Michael Hargiss <mhargiss@instructure.com>
Product-Review: Derrick Hathaway <derrick@instructure.com>