Make minical not be read by NVDA after opening color picker
Essentially this is a bug in the way NVDA/Firefox work.  The color
picker uses react-modal. react-modal does a aria-hidden toggle on the
application div when opening closing.  NVDA will then not respect the
aria-hidden that is on the minical (even though it should).

I filed https://github.com/nvaccess/nvda/issues/6053 against NVDA to
have them look into it.  But for now, doing a redraw forcing the SR
to re-read the page tends to work.  This takes a helper function that
was in use in Files, converts it to using ES2015, and puts it in a
shared area.

There are no tests associated with this commit because we have no way
to test specific screenreader behavior like this.

closes CNVS-29813

Test Plan:
  - Using NVDA/Firefox go to the Calendar (/calendar)
  - Navigate to the list of calendars (the minical shouldn't have been read.)
  - Go to the button to open the color picker, then close the color picker
  - Navigate back through the page, the minical should likewise still
    not be read.

Change-Id: If62a194667f2ebf589ee639ec8a66b7e39863f2b
Reviewed-on: https://gerrit.instructure.com/81931
Tested-by: Jenkins
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Deepeeca Soundarrajan <dsoundarrajan@instructure.com>
Product-Review: Clay Diffrient <cdiffrient@instructure.com>