Make minical not be read by NVDA after opening calendar feeds
Essentially this is a bug in the way that NVDA and/or Firefox
works. While the aria-hidden is true on the minical and should
always be respected, a bug (https://github.com/nvaccess/nvda/issues/6053)
causes changes to the parents aria-hidden to unhide the aria-hidden
stuff in its children :(

There are no tests associated with this commit because there is no
real way to test specific screenreader behavior.

closes CNVS-30258

Test Plan:
  - Using NVDA/Firefox go to the Calendar (/calendar)
  - Navigate to the calendar feed button (the minical shouldn't have been read.)
  - Click the button, then close the modal
  - Navigate back through the page the minical should still not be read.

Change-Id: I48f8698c41b81c8b3b8b128be8b7b19fb710454a
Reviewed-on: https://gerrit.instructure.com/84093
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Jahnavi Yetukuri <jyetukuri@instructure.com>
Product-Review: Clay Diffrient <cdiffrient@instructure.com>