fix syllabus description not showing in rcs editor
state is being maintained in a data attribute of the containing div,
and is only set in the text area when showing the edit form. after
the first edit is was being set on a stale node.

fixes CNVS-30322

test plan:
- enable the remote editor with sidebar feature flag
- go to syllabus
- click edit
- made a change
- click save
- click edit again
  - the editor should be populated with the right content, not empty

Change-Id: I568be0815265d28728c11687068d73e723571a25
Reviewed-on: https://gerrit.instructure.com/84562
Tested-by: Jenkins
Reviewed-by: Simon Williams <simon@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Simon Williams <simon@instructure.com>