upgrade canvas_quizzes npm deps so it is ready for node 6
closes: CNVS-30448

in order to get node 6 to build and run on jenkins,
we have to upgrade grunt-contrib-jasmine to work
around this bug: https://github.com/Medium/phantomjs/issues/430

and we have to upgrade gulp-sass so the node-sass
version it asks for is node-sass@3.8.0 to get it to
actually build on jenkins. as in, so there
is a precompiled binary for it that is compiled
against the v8 in node 6. (there are none for 0.9.3
against node 6, since node 6 didn't exist when 0.9.3
was released).  This is necessary because
the gcc on jenkins is too old to compile node-sass
from source.

test plan:

* quzzes jenkins build should pass
* click around in quiz stats and quiz log auditing
  to make sure things load

Change-Id: Ic56fae34c52aab5d9b7971fe871d04fcc0738206
Reviewed-on: https://gerrit.instructure.com/85389
Tested-by: Jenkins
Reviewed-by: Clay Diffrient <cdiffrient@instructure.com>
QA-Review: Michael Hargiss <mhargiss@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>