fix incorrect canvas.externaltool.url substitution
Fixes #880  Github issue

fixes PLAT-1770

test plan:
add a tool installed at the account level with the
"Canvas.externalTool.url" variable expansion

the substituted value should be the tools context not the launch
context i.e. if the tool is launched from a course below the
account it is installed on it should be a url for the account.

also the returned url should be the correct api endpoint for the
tool

Change-Id: Ibb166240c96fb01b31bc2d5055b67f16e107878c
Reviewed-on: https://gerrit.instructure.com/88539
Tested-by: Jenkins
Reviewed-by: Weston Dransfield <wdransfield@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: Nathan Mills <nathanm@instructure.com>