don't blow up when address book service misbehaves
the address book service, when enabled, will fulfill the
known_in_context request which provides the list of users. however, as
of now, the address book service is still in development and can return
extra users that _aren't_ in context for that call. this will be fixed
in the address book service, but in the meantime let canvas handle the
service's errors without blowing up itself, so that we can still test
integration with other parts of the address book.

test-plan:
 - have the address book service running and canvas configured to talk
   to it
 - have data in the address book service so that the logged in user
   knows at least one other user
 - have a course in which the logged in user is enrolled but no other
   "known" user is enrolled
 - as the logged in user, visit the inbox and open the address book in
   the recipients input for a new conversation.
 - should not get an error

Change-Id: I9e815aee0e36fc49ef1d6a46d363b651ab54e9b7
Reviewed-on: https://gerrit.instructure.com/91910
Tested-by: Jenkins
Reviewed-by: Andrew Huff <ahuff@instructure.com>
QA-Review: Cemal Aktas <caktas@instructure.com>
Product-Review: Jacob Fugal <jacob@instructure.com>