Clarify "rich editor" and "link selection" api docs
Documentation was innacurrate surrounding the "text" query parameter.
Before this fix, the field was marked as optional; it is now required.
Furthermore, it was implied that one could highlight text and replace the
returned url would be used as a hyperlink for that text; now it states that
the highlighted text will be replaced. Another patchset will update these
same endpoints to state that resource selection is being deprecated in
place of content-item. This fix does not address that.

Test plan:
1) run localhost Canvas
2) build the API docs (docker-compose run --rm web bundle exec rake
doc:api)
3) navigate to <your host url>/doc/api/file.link_selection_tools.html and
<your host url>//doc/api/file.editor_button_tools.html
4) check for spelling, grammar, accuracy of information and that the new
link works

Fixes PLAT-1920

Change-Id: Ib2b563d3902935441f76f582112a99dd53d8d56d
Reviewed-on: https://gerrit.instructure.com/93869
Reviewed-by: Weston Dransfield <wdransfield@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Reviewed-by: Andrew Butterfield <abutterfield@instructure.com>
Tested-by: Jenkins
Product-Review: Andrew Butterfield <abutterfield@instructure.com>