Add "SIS syncing" checkbox to account admin settings
This only adds the SIS syncing checkbox. It does NOT implement all
the logic related to the controls under it, as those are coming
in later stories.

The majority of the front-end changes represent a refactoring of the
existing ERB into partials. I recognize the official policy is to not
add any new ERB files, but rather to implement new UI functionality
in react. This is not new functionality, it simply makes the existing
markup easier to manage, easier to see how the various feature flags
and account settings impact what is rendered.

Fixes: SIS-2571

Test plan:
- With new SIS integrations feature DISABLED:
  - no "SIS syncing" checkbox on the settings page
- With new SIS integrations feature ENABLED:
  - "SIS syncing" checkbox is present in the settings page
  - other syncing settings ('Post grades to SIS', lock to sub-accounts)
    are immediately made visible when "SIS syncing" is checked,
    immediately hidden when unchecked.

Change-Id: Ib51e0b3ab405e84fdc54fac55dcfbf73cc4e0439
Reviewed-on: https://gerrit.instructure.com/94748
Tested-by: Jenkins
Reviewed-by: Stewie aka Nicholas Stewart <nstewart@instructure.com>
QA-Review: Adam Phillipps <adam@instructure.com>
Product-Review: Oxana