fix a case where dateJS could get into infinite loop
basically, it thought it was calling back to the 
native date.prototype.toString but it was really
just calling itself, so it would keep doing that 
infinitely.

test plan:
this didn’t happen very often but sometimes the
calendar page would throw an error on load because
the it ran out of js stack size. so it is pretty
hard to think of a definitive test of where it didn’t
work before but if you 
open calendar page and files page and they work,
this should be good

Change-Id: I1beec8347b46582761912a2137640bfd3fe6b9e2
Reviewed-on: https://gerrit.instructure.com/95845
Reviewed-by: Simon Williams <simon@instructure.com>
Tested-by: Jenkins
Product-Review: Ryan Shaw <ryan@instructure.com>
QA-Review: Ryan Shaw <ryan@instructure.com>