Make finding external tools by domain case insensitive
Currently, if the case of the domain does not match the case of the host
for an LTI launch URL, Canvas is unable to find the tool. Eg.: if the
lti_launch_url that is clicked is 'https://tool.host.com', but the tool is
configured with a domain of "Host.com", Canvas throws "Couldn't find valid
settings for this link". This patch downcases both the host and the
domain before tying to find a match, effectively making tha match case
insensitive.

Test plan:
1) configure an LTI tool with only a domain (ie, no launch url). Have at
least one upper case letter in the domain (eg. Host.com)
2) add the tool as a module item in a course, making sure to specify a
launch url that matches the domain, but is all lower-case
3) try to launch the tool
4) Canvas should not throw the error "Couldn't find valid settings for this
 link"

Fixes PLAT-2008

Change-Id: Iae4a806cee45779ce94f27347519eb5a37c92514
Reviewed-on: https://gerrit.instructure.com/96381
Reviewed-by: Andrew Butterfield <abutterfield@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Tested-by: Jenkins
Product-Review: Jesse Poulos <jpoulos@instructure.com>