SIS Setting: Rename generic SIS to "Your SIS" Sync option
Fixes: SIS-2576

Notes: The gergich warning about adding tests is
being ignored because we are only adding a setting
to accounts which is tested else where i.e. the UI
where it is actually used and tested thoroughly.
This has been the pattern of the other account
settings we've added in the past.

Test plan:
- Confirm new UI changes are present
- Add new "SIS" name
- Confirm the setting persists when you save

Change-Id: I1ea8974cc55c95e7eefe6249766e9f06f6949cf4
Reviewed-on: https://gerrit.instructure.com/96915
Reviewed-by: Stewie aka Nicholas Stewart <nstewart@instructure.com>
Tested-by: Jenkins
QA-Review: Steven Shepherd <sshepherd@instructure.com>
Product-Review: Oxana