spec: add gb test for closed grading period
Right now we have to use CSS in the spec file to
avoid putting the rspec matcher in the page object
file.  In the future, we plan to write a custom selenium
function 'contains_css' for use in page object files,
but for now this is the best solution.

Change-Id: I575246fd6613db1c4ecaf0f7f8949ced0f8471a6
Reviewed-on: https://gerrit.instructure.com/96998
Tested-by: Jenkins
Reviewed-by: KC Naegle <knaegle@instructure.com>
Product-Review: Gentry Beckmann <gbeckmann@instructure.com>
QA-Review: Gentry Beckmann <gbeckmann@instructure.com>