refactor grading_standards tab containers
This will make it more obvious that tabContainer.jsx
is for course by renaming the files to CourseTabContainer.
This style of rename is present in other files now as well.

This patchset also adds specs to ensure that not only is jQuery-UI's
tabs() function is called, but that the corresponding DOM nodes
are decorated with the correct classes and styles.

closes: CNVS-34416

test plan:
 - ensure the grading_standards page renders correctly with
   grading periods disabled/enabled on both course and
   accounts grading_standards.

Change-Id: I125f3c3fd770972cc42f572529066e1e2ca4e40d
Reviewed-on: https://gerrit.instructure.com/99487
Tested-by: Jenkins
Reviewed-by: Spencer Olson <solson@instructure.com>
Reviewed-by: Keith T. Garner <kgarner@instructure.com>
QA-Review: Anju Reddy <areddy@instructure.com>
Product-Review: Keith T. Garner <kgarner@instructure.com>