add rubocop rules for rspec nesting and http arguments
This disables the HttpPositionalArguments rubocop rule, which is invalid
for versions of Ruby prior to Ruby 5. This also sets the limit for
nested example groups in RSpec to 3, which is a more reasonable limit
than the default of 2.

test plan:
* verify Jenkins passes

Change-Id: I8f2badb369d6932923aa4c096c9b9014be19805a
Reviewed-on: https://gerrit.instructure.com/100198
Tested-by: Jenkins
Reviewed-by: Rob Orton <rob@instructure.com>
QA-Review: Jeremy Neander <jneander@instructure.com>
Product-Review: Jeremy Neander <jneander@instructure.com>