Fix sticky header issues in IE
Fixes: FALCOR-378

Note for CR:
Sticky positioning CSS has to be applied to the outermost dashboard
header DIV for it to work (it needs to be a sibling of the main dashboard
content).

Unfortunately, the dashboard header DIV is also the mount point for the
<DashboardHeader /> component, and I couldn't figure out how to add
classes to it based on what dashboard view was showing.

This is why I'm adding classes to the body :(

Test plan:

Log into Canvas as a student. Make sure the planner
is turned on in accounts settings first.

Chrome, Firefox, Safari:
- All dashboard headers (Cards, List, Recent Activity) should be sticky

IE:
- The dashboard header will only be sticky for Planner
- You should no longer see the bug described in the JIRA ticket

Change-Id: Ie1bcf30567ff91413ee0b569c9ed434d1fd1320f
Reviewed-on: https://gerrit.instructure.com/116620
Tested-by: Jenkins
Reviewed-by: Steven Burnett <sburnett@instructure.com>
QA-Review: Dan Sasaki <dsasaki@instructure.com>
Product-Review: Steven Burnett <sburnett@instructure.com>