spec: replace column rendering spec
The code covered by this spec was mostly covered in two other
specs which cover the only consumers of this code. This moves the
assertions from this spec to those, which will allow this
function to be refactored safely in subsequent changes.

refs CNVS-38346

test plan:
* ensure Jenkins passes

Change-Id: Ie2075296002e93bf212c6da052dc9d4fe5d08a3f
Reviewed-on: https://gerrit.instructure.com/121509
Reviewed-by: Gary Mei <gmei@instructure.com>
Tested-by: Jenkins
Product-Review: Jeremy Neander <jneander@instructure.com>
QA-Review: Jeremy Neander <jneander@instructure.com>