Fix High Contrast for MS edge 15
Fixes: CNVS-38996

If this style rule is included, the entire page
Crashes in Edge 15. Not a server error, not a JS
error, Edge itself crashes.

UI-Devs: if you still want to have this different
box-shadow in High Contrast mode, you could maybe
toy with why it caused the page to crash or do it in a way
That it doesn’t apply to Edge 15. But deleting it was
The easiest way I found to fix the problem

Test plan:
* in edge 15 (you can go to modern.ie and launch a live edge 15 from 
  your normal browser for free)
* log in and set your user settings to use high contrast
* go to any page
* edge should not crash and the page should display correctly

Change-Id: I2df0da7e328f5200a9fe1fa9235d28bbb561a10a
Reviewed-on: https://gerrit.instructure.com/127216
Tested-by: Jenkins
Reviewed-by: brian kirkby <bkirkby@instructure.com>
Reviewed-by: Chris Hart <chart@instructure.com>
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Product-Review: Ryan Shaw <ryan@instructure.com>