Fix focus on duplicating/removing module item.
On duplicate, focus will go to the link to the newly created item.

On delete, focus will go to the previous item's link (if the
previous item is a header, then it will go to the cog instead).
If the first item is deleted, the current behavior of going to the
cog of the module is preserved.

Fixes COMMS-510.

Test Plan:
* Have a module with some items.
* On duplicate, focus should go to the link on the newly created
  item.
* On deleting something just below a header, the focus should go
  on the header's cog.
* On deleting something not just below a header, focus should
  go on the previous item's link.

Change-Id: Ib3686b95f5dbdf795cc8ddaec0bdf0ba435b163c
Reviewed-on: https://gerrit.instructure.com/132672
Tested-by: Jenkins
Reviewed-by: Felix Milea-Ciobanu <fmileaciobanu@instructure.com>
QA-Review: Steven Burnett <sburnett@instructure.com>
Product-Review: Venk Natarajan <vnatarajan@instructure.com>