Mostly revert "Enable viewing of CalendarEvents cross shard"
Keep the api change though, to return asset strings relative
to the current shard.

And the course enrollments forcing the query to be relative
to the correct shard.

Also add a migration to drop the table that may or may not exist.

This reverts commit 7a4aac73e25bf0cff0bb0cb034d1d832537e0923.

Change-Id: I5f6d9d6499fe507f98198fe546e31d8059933b76
Reviewed-on: https://gerrit.instructure.com/133960
Tested-by: Jenkins
QA-Review: Jeremy Putnam <jeremyp@instructure.com>
Reviewed-by: Tyler Pickett <tpickett@instructure.com>
Product-Review: Cody Cutrer <cody@instructure.com>